# perf(test): pre-spawn neovim instances