Functional tests call `clear()` before each test, which spawns a new neovim instance, and start using it immediately after. This forces each test to wait until the newly spawned neovim is ready.

Added a queue of 10 neovim instances that are spawned before they are needed. When `clear()` is called, the oldest instance in the queue is used. This makes it more likely that the instance is ready by the time it is needed. Queue is refilled when instances are removed from it.

This reduces total time taken by functional tests.

Note:

Since instances are created before the tests that use them, we can't use `_G._nvim_test_id` from https://github.com/neovim/neovim/pull/8519 in e.g. `test/functional/core/log_spec.lua`, since we can't know which test will use which instance. New `_G._nvim_session_id` is introduced, which is set when the session is set.

Other places like making a temporary directory of printing the name of the test still use `_G._nvim_test_id`.

Also some of the calls to `clear()` have arguments, but since we can't predict which arguments will be passed, the optimization doesn't apply to them.

### Results

|| current | new |
|-|-|-|
ubuntu asan clang functionaltest | 14m 54s | 12m 38s |
ubuntu tsan clang functionaltest | 15m 56s | 14m 41s |
ubuntu release gcc functionaltest | 6m 30s | 6m 15s |
macos intel clang functionaltest | 9m 36s | 7m 44s |
macos arm clang functionaltest | 6m 29s | 5m 40s |
ubuntu puc-lua gcc functionaltest | 6m 46s | 6m 28s |
windows / windows (functional) | 11m 35s | 8m 34s |


<details>

<summary>data</summary>

current:
1. https://github.com/vanaigr/neovim/actions/runs/13501048829
2. https://github.com/vanaigr/neovim/actions/runs/13512336529
3. https://github.com/neovim/neovim/actions/runs/13549176090
4. https://github.com/neovim/neovim/actions/runs/13548758242
5. https://github.com/neovim/neovim/actions/runs/13548725467

||1|2|3|4|5|
|-|-|-|-|-|-|
| ubuntu asan clang functionaltest | 15m 26s | 14m 29s | 14m 59s | 14m 32s | 15m 2s |
| ubuntu tsan clang functionaltest | 15m 15s | 15m 57s | 16m 19s | 17m 3s | 15m 5s |
| ubuntu release gcc functionaltest | 6m 35s | 6m 19s | 6m 38s | 6m 18s | 6m 41s |
| macos intel clang functionaltest | 11m 25s | 8m 37s | 8m 51s | 9m 2s | 10m 5s |
| macos arm clang functionaltest | 6m 28s | 6m 33s | 6m 51s | 6m 19s | 6m 12s |
| ubuntu puc-lua gcc functionaltest | 6m 39s | 6m 41s | 6m 50s | 6m 46s | 6m 56s |
| windows / windows (functional) | 12m 24s | 11m 42s | 11m 20s | 11m 20s | 11m 10s |


new:
1. https://github.com/vanaigr/neovim/actions/runs/13550052710
2. https://github.com/vanaigr/neovim/actions/runs/13549525879
3. https://github.com/vanaigr/neovim/actions/runs/13539200286

||1|2|3|
|-|-|-|-|
| ubuntu asan clang functionaltest | 12m 25s | 12m 53s | 12m 36s |
| ubuntu tsan clang functionaltest | 15m 6s | 14m 14s | 14m 42s |
| ubuntu release gcc functionaltest | 6m 28s | 6m 11s | 6m 7s |
| macos intel clang functionaltest | 7m 8s | 8m 32s | 7m 31s |
| macos arm clang functionaltest | 5m 57s | 5m 23s | 5m 41s |
| ubuntu puc-lua gcc functionaltest | 6m 17s | 6m 41s | 6m 25s |
| windows / windows (functional) | 8m 36s | 8m 40s | 8m 25s |


```python
def average_times(time_lists):
    averaged_times = []
    for times in time_lists:
        total_seconds = sum(m * 60 + s for m, s in times)
        avg_seconds = round(total_seconds / len(times))
        avg_minutes, avg_seconds = divmod(avg_seconds, 60)
        averaged_times.append(f"{avg_minutes}m {avg_seconds}s")
    return averaged_times

time_data = [
    # [(12, 25), (12, 53), (12, 36)],
    ...
]

print(average_times(time_data))
```

</details>