@vanaigr 2025-02-26
test `fileinfo does not overwrite echo message vim-patch:8.2.4156` ([link](https://github.com/neovim/neovim/blob/0f24b0826a27b7868a3aacc25199787e7453d4cc/test/functional/legacy/messages_spec.lua#L666)) started failing because before, I assume, all 3 commands were executed at the same time because it took a long time for the test instance to become active.

Moved these two commands:
https://github.com/neovim/neovim/blob/0f24b0826a27b7868a3aacc25199787e7453d4cc/test/functional/legacy/messages_spec.lua#L683-L684

together with the previous commands. Otherwise the cursor ends up on the last charcater.