@vanaigr 2025-03-02
> Converted to draft until I fix the windows test. The same test failed twice

Looks like the same thing happens as in the first failing test: neovim is ready too soon, and the test doesn't expect this.

https://github.com/neovim/neovim/blob/188ec19894e4d3abadec7b9a43fecf62dd77b376/test/functional/ui/diff_spec.lua#L1559-L1565

The files are read (by `args` command) before they are written. Executing commands one by one leads to the same outcome.

Original oldtest sets `autoread` before writing the files:
https://github.com/neovim/neovim/blob/188ec19894e4d3abadec7b9a43fecf62dd77b376/test/old/testdir/test_diffmode.vim#L1866-L1871

though setting it doesn't reload the files for me on windows.