@vanaigr 2025-03-11
I tried to reproduce the issue locally but couldn't.

After multiple runs, it seems that the files are not reloaded when they should. Even waiting for 100ms didn't help. But forcing reloading with `edit!` does help.

https://github.com/neovim/neovim/commit/647cf50f588e379f7cd85008674a112fc3d707d9