@vanaigr 2025-07-05
I don't have time to work on it. It appears it still fails on windows with diff issues, and debugging would be very time consuming