@justinmk 2025-07-06
> Since instances are created before the tests that use them, we can't use `_G._nvim_test_id` from #8519 in e.g. `test/functional/core/log_spec.lua`, since we can't know which test will use which instance. New `_G._nvim_session_id` is introduced, which is set when the session is set.

I think that's fixable without introducing `_G._nvim_session_id`. In the pooled nvim instance, call `serverstop(...) ... serverstart(_nvim_test_id)` in a single RPC call. 
That will close the channel, but should be able to reopen with the new address.
That should trigger [set_vservername](https://github.com/neovim/neovim/blob/4f141dca8c888d62b332f31fd51fa3e9b2e6cb89/src/nvim/msgpack_rpc/server.c#L203-L204), then the logs [should use](https://github.com/neovim/neovim/blob/4f141dca8c888d62b332f31fd51fa3e9b2e6cb89/src/nvim/log.c#L341) the updated "primary name" of the Nvim sever.

If that doesn't work, we can provide some way to actually set the Nvim "primary name". That will avoid `_G._nvim_session_id`.