@vanaigr 2025-07-06
> > it may start failing randomly again in the future
> 
> "Randomly" because of a future mistake in the tests?

Diff tests fail, and I don't know the reason. It has something to do with Windows and file change detection. 
I thought I found the issue, but then the test failed again.

And this might happen again. I would think I found the fix, and it would fail again (e.g. after the PR got merged)



> What if we disabled this for certain test files (i.e. certain test files never use the "queue"), is that hard to do?

Yes, we can disable it