@vanaigr 2025-07-06
I can't reopen the PR though. See: https://github.com/isaacs/github/issues/361