@vanaigr 2025-07-06
We can decide what to do with `_nvim_session_id`. Then I'll add option to disable pooling for files. And then open a new PR