@justinmk 2025-07-06
> Removing conditional log name regen fixes this:

you mean removing the `if (regen)`? We may need to store the old value of `get_vim_var_str(VV_SEND_SERVER)` and then compare it, like:

    const char *serv = path_tail(get_vim_var_str(VV_SEND_SERVER));
    bool regen = ui || name[0] == NUL || name[0] == '?' || !strequal(serv_name, serv);